Fixed bug where an error in the deactivate lua callback would recurse infinitely
GTA far clip distance is 800 by default

Cars finally no longer drift when parked (they do drift when still and not
parked, though) It ought to be possible to use this code when the cars are
still as well as when parked, but there are problems with the wheels turning
slowly on the spot.

accelerate + handbrake is now treated as wheel spin, i.e. for powersliding without losing speed

fixed bug where brake curve was used from object instead of class

speed boost is now more realistic, uses direction of motion instead of direction of facing

tweaked seat collision mesh to allow driving onto curbs

test.lua is now at an angle